erl_driver.h: Enlarge types in driver output functions

+++ erts/emulator/beam/erl_debug.h
@@ -1,7 +1,7 @@
 /*
  * %CopyrightBegin%
  *
- * Copyright Ericsson AB 2004-2010. All Rights Reserved.
+ * Copyright Ericsson AB 2004-2011. All Rights Reserved.
  *
  * The contents of this file are subject to the Erlang Public License,
  * Version 1.1, (the "License"); you may not use this file except in
@@ -51,7 +51,7 @@
 
 extern Uint32 verbose;
 
-void upp(byte*, int);
+void upp(byte*, size_t);
 void pat(Eterm);
 void pinfo(void);
 void pp(Process*);
